refactor decompression code into reference and platform specific Apply some Xtensa acceleration code changes
